--- web/codex/map_details.php
@@ -1,30 +1,0 @@
-<?php
-require_once($_SERVER['DOCUMENT_ROOT'] . "/bullet_hell/web/src/php/utils.php");
-require_once($_SERVER['DOCUMENT_ROOT'] . "/bullet_hell/web/src/php/links.php");
-if (!is_logged_in()) {
-    header("Location: ../login/login.php");
-}
-$mapId = $_GET['id'];
-?>
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Maps</title>
-    <link rel="stylesheet" href="src/style.css">
-</head>
-
-<body>
-    <?php require_once($_SERVER['DOCUMENT_ROOT'] . "/bullet_hell/web/src/php/header.php") ?>
-    <script src="src/map_details.js" type="module"></script>
-    <div class="map-details">
-
-    </div>
-    <div class="map-weapons">
-
-    </div>
-</body>
-
-</html>
