+++ web/codex/src/carousel_style.css
@@ -1,7 +1,7 @@
 html, body {
     height: 100%;
     margin: 0; 
-    overflow: hidden;
+    font-family: 'Pixelmix', sans-serif;
 }
 
 #image-carousel {
@@ -10,7 +10,18 @@ html, body {
     position: relative;
     overflow: hidden;
 }
-
+#carousel-wrapper {
+    width: 100%;       
+    max-width: 170vh;   
+    aspect-ratio: 16 / 9;
+    position: relative;  
+}
+#image-carousel {
+    width: 100%;
+    height: 100%;
+    position: relative;
+    overflow: hidden;
+}
 .carousel-item {
     position: absolute;
     width: 100%;
@@ -26,13 +37,6 @@ html, body {
     object-fit: contain;
     flex-grow: 1;
 }
-
-.carousel-item img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
 .carousel-item.active {
     display: block;
     transform: translateX(0);
@@ -64,7 +68,7 @@ html, body {
     height: 300px;
     background: linear-gradient(to bottom, rgba(0,0,0,0), rgba(0,0,0,1));
     color: white;
-    padding: 20px;
+    padding: 130px 20px 0px 20px;
     box-sizing: border-box;
     z-index: 5;
     text-align: right;
